added Payment_detail model , migration and relationed with order_total

+++ database/migrations/2022_04_28_142852_create_shopping_sessions_table.php
@@ -17,7 +17,7 @@ class CreateShoppingSessionsTable extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id')->nullable();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
-            $table->bigInteger('total');
+            $table->float('total', 6, 2);
             $table->timestamps();
         });
     }
